+++ src/packer/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to Packer will be documented in this file.
 
+## [1.1] - 2026-04-22
+
+* 0e3e13a - Signature calculating improved
+* 2af2ef4 - The value set by the Packer::SetSalt() method affects encryption
+* 30efd4b - An InvalidArgumentException is thrown when the variable cannot be JSON-encoded
+* f725ea1 - [Security] Raw binary key
+
 ## [1.0] - 2026-04-21
 
 First release as a standalone package.